Doc
Documentation de code utilisant doxygen

+++ view/delete_product.php
@@ -1,17 +1,28 @@
 <?php
-$host = 'localhost';
-$user = 'Ecommerce';
-$password = 'Stage2023*';
-$database = 'E-commerce';
+/**
+ * @file delete_product.php
+ * @brief Script pour supprimer un produit de la base de données.
+ */
 
-$connection = mysqli_connect($host, $user, $password, $database);
+$host = 'localhost'; /**< Adresse du serveur de base de données. */
+$user = 'Ecommerce'; /**< Nom d'utilisateur de la base de données. */
+$password = 'Stage2023*'; /**< Mot de passe de la base de données. */
+$database = 'E-commerce'; /**< Nom de la base de données. */
 
+$connection = mysqli_connect($host, $user, $password, $database); /**< Connexion à la base de données. */
+
+// Vérification de la connexion à la base de données
 if (!$connection) {
     die("Connection failed: " . mysqli_connect_error());
 }
 
-if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-    $product_id = $_POST['product_id'];
+/**
+ * @brief Supprime l'image associée au produit de la base de données.
+ * @param int $product_id Identifiant du produit.
+ */
+function deleteProductImage($product_id)
+{
+    global $connection;
 
     $imageQuery = "SELECT image FROM products WHERE product_id = '$product_id'";
     $imageResult = mysqli_query($connection, $imageQuery);
@@ -25,10 +36,18 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             unlink($imagePath);
         }
     }
+}
+
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    $product_id = $_POST['product_id'];
+
+    // Suppression de l'image associée au produit
+    deleteProductImage($product_id);
 
     $query = "DELETE FROM products WHERE product_id = '$product_id'";
 
     if (mysqli_query($connection, $query)) {
+        // Redirection vers la page d'administration des produits
         header("Location: admin.php?tab=view_products");
         exit();
     } else {
@@ -36,4 +55,5 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     }
 }
 
-mysqli_close($connection);
+mysqli_close($connection); /**< Fermeture de la connexion à la base de données. */
+?>
